Enqueue events, and only propagate on next microtick
This is still just an experiment.  Delaying all events by a microtick
does help with the race condition of pushing events in the same tick.
If consumers are far away in time and space from the point of event
push, there will always be the potential for race conditions between
pushing events and attaching consumers.