Add transduce() for applying transducers to event streams.
Unfortunately, I had to resort to a stateful approach for threading
event times through the transducer machinery by storing the time on
the transformer.  I couldn't find a reliable way to do it otherwise,
due to a difference in the transducer prototype vs. most.js's internal
protocol.  I think this is ok, since the statefulness is isolated and
completely out of reach of app-level code.