Use native Promise
The removes when.js's es6 shim in favor of using native Promise.
Environments that don't have Promise will need to use a polyfill.

There are probably a few spots that will incur a small perf hit since
native Promise can be significantly slower.  One of those was in
mergeConcurrently, and I was able to refactor it slightly to avoid the
hit in cases where disposal doesn't return a promise.  If it's a
problem, apps can use a fast polyfill like creed, when, or bluebird.