Add ES7 interop (#254)
* Add initial ES7 Observable interop

This is the start of minimal interoperability support for things that
provide [Symbol.observable].  It will not be a full implementation of
the proposed spec, but will allow compatibility with other libs that
produce and consume Observable-compatible things.

* Support second order observables

mergeMapConcurrently and switch now support coercing inner
observables.  E.g. A most stream chain()ed using a function that
returns another lib's Observable will work.

* Add observable support for continueWith and recoverWith

* Ensure subscribe returns subscription

* Add getObservable tests

* Add unit tests for Subscription and SubscribeObserver

* Make from() identity for mostjs streams

* Add fromObservable tests

* Add test to verify ES7 Observable API exists. Switch to assert

* Start observable interop docs

* More docs

* Remove implicit coercion

* Finish ES7 Observable interop docs section

* Add subscribe docs. Tweak ES7 interop docs

* Skip non-runnable examples. Remove coerceToStream

* Add guidance/info on forEach vs subscribe

* Fix syntax

* Capitalize types