Fixes most.d.ts merge, mergeArray definition
The current definition restricts only single generic type of A. merge parameters should have multiple generics.

Right now it is limited to to up to 5 types.

Note: Most does not restrict this `most.merge(stream)` in javascript. but this will fail in typescript since we typed it to either 2 or more types.